+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.36.0
 
 * Add support for Apache Kafka 3.4.1 and 3.5.0, and remove support for 3.3.1 and 3.3.2
+* Enable SCRAM-SHA authentication in KRaft mode (supported in Apache Kafka 3.5.0 and newer)
 * Update Kafka Exporter to [1.7.0](https://github.com/danielqsj/kafka_exporter/releases/tag/v1.7.0)
 * Added support for Kafka Exporter topic exclude and consumer group exclude parameters
 * Update Kaniko container builder to 1.11.0